refactor: update DTOs to match weight type changes - Change peso type from Integer to Float in DTOs - Add status field to HistoricoPesoResponse - Update HistoricoPesoMapper to include status field

+++ src/main/java/dev/caio/fitsy/dto/response/UserInfoResponse.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 public record UserInfoResponse(Integer status,
                                String mensagem,
                                Float peso,
-                               Float altura,
+                               Integer altura,
                                Sexo sexo,
                                LocalDate data_nascimento) {
 }
